pyproject.toml
[tool.poetry]
name = "todolist"
version = "0.1.0"
description = "skypro final project"
authors = ["frame <frame.boo@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.1.4"
django-environ = "^0.9.0"
psycopg2-binary = "^2.9.5"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
social-auth-app-django = "^5.0.0"
django-filter = "^22.1"
marshmallow = "^3.19.0"
marshmallow-dataclass = "^8.5.11"
requests = "^2.28.2"
pytest-django = "^4.5.2"
factory-boy = "^3.2.1"
pytest-factoryboy = "^2.5.1"


[tool.poetry.group.dev.dependencies]
pre-commit = "^2.20.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
